requirements.txt
fastapi~=0.115.12
requests~=2.32.3
python-dotenv~=1.1.0
transformers~=4.51.3
librosa~=0.11.0
torch~=2.2.2
uvicorn
safetensors
anthropic~=0.52.0
python-multipart

pydantic~=2.11.4

